feat: run examples in seprate processes
